ロジックをuseHooksに切り出し、カウントのロジックを変更
