+++ src/App.js
@@ -1,79 +1,25 @@
-import { useState } from 'react';
+// import { useState } from 'react';
 
 import { InputTodoArea } from './components/InputTodoArea';
+import { useHooks } from './hooks/hooks'
 import './App.css';
 
 function App() {
-  const [todoItems, setTodoItems] = useState([]);
-  const [incompleteCount, setIncompleteCount] = useState(0);
-  const [completeCount, setCompleteCount] = useState(0);
-  const [text, setText] = useState('');
-  const [editTodoText, setEditTodoText] = useState('');
-
-  const onChangeText = (event) => setText(event.target.value);
-
-  const onChangeTodoText = (event) => setEditTodoText(event.target.value);
-
-  const onClickAdd = () => {
-    const newTodoItems = [...todoItems];
-    const addTodoItem = {
-      text,
-      edited: false,
-      checked: false
-    };
-    setTodoItems([...newTodoItems, addTodoItem]);
-    setText('');
-    setIncompleteCount(incompleteCount + 1);
-  };
-
-  const onClickEditForm = (index) => {
-    const todoTemp = [...todoItems];
-    const editTodoItem = todoItems[index];
-    editTodoItem.edited = true;
-    todoTemp[index] = editTodoItem;
-    setEditTodoText(editTodoItem.text);
-    setTodoItems(todoTemp);
-  };
-
-  const editTodo = (index) => {
-    const todoTemp = [...todoItems];
-    todoTemp[index].text = editTodoText;
-    todoTemp[index].edited = false;
-
-    setEditTodoText('');
-    setTodoItems(todoTemp);
-  }
-
-  const onClickDelete = index => {
-    if (window.confirm('本当に削除してもよろしいでしょうか?')) {
-      const todoTemp = [...todoItems];
-      const deleteTodo = todoItems[index];
-      if (deleteTodo.checked) {
-        setCompleteCount(completeCount - 1);
-      } else {
-        setIncompleteCount(incompleteCount - 1);
-      }
-      todoTemp.splice(index, 1);
-      setTodoItems(todoTemp);
-    }
-  }
-
-  const onChangeCheck = (index) => {
-    const todoTemp = [...todoItems];
-    const checkedTodoItem = todoItems[index];
-    checkedTodoItem.checked = !checkedTodoItem.checked;
-    todoTemp[index] = checkedTodoItem;
-    setTodoItems(todoTemp);
-
-    // 表示するタスクの数を変更
-    if (checkedTodoItem.checked) {
-      setIncompleteCount(incompleteCount - 1);
-      setCompleteCount(completeCount + 1);
-    } else {
-      setIncompleteCount(incompleteCount + 1);
-      setCompleteCount(completeCount - 1);
-    }
-  }
+  const {
+    todoItems,
+    text,
+    editTodoText,
+    onChangeText,
+    onChangeTodoText,
+    onClickAdd,
+    onClickEditForm,
+    editTodo,
+    onClickDelete,
+    onChangeCheck
+  } = useHooks();
+  const allCount = todoItems.length;
+  const completeCount = todoItems.filter(item => item.checked).length
+  const incompleteCount = allCount - completeCount;
 
   return (
     <>
